+++ packages/components/src/fluent/index.ts
@@ -1,4 +1,5 @@
 export * from "../shared/definition";
+export { fluentProjectionProvider } from "../GikComponentDeclarative";
 export * from "./authoring";
 export * from "./FluentButtons";
 export * from "./FluentBasicControls";
@@ -7,6 +8,7 @@ export * from "./FluentDialog";
 export * from "./FluentDisplayControls";
 export * from "./FluentInputs";
 export * from "./FluentLayoutControls";
+export * from "./FluentStack";
 export * from "./FluentText";
 export * from "./FluentToolbar";
 export * from "./registry";