+++ packages/components/src/security/index.ts
@@ -1,4 +1,5 @@
 export * from "../shared/definition";
+export { securityProjectionProvider } from "../GikComponentDeclarative";
 export * from "./authoring";
 export * from "./registry";
 export * from "./attack-path";